Remove support for the xCal format (RFC 6321).
As far as we know, no mainstream clients use this format,
and removing support for it reduces code bloat.
